# api-spec.yaml
openapi: 3.0.3
info:
  title: Order Management Microservices API
  description: |
    Comprehensive API specification for Order Management System with SAGA pattern implementation.
    This includes Order, Payment, Inventory, and Notification microservices.
  version: 1.0.0
  contact:
    name: API Support
    email: support@orderms.com
servers:
  - url: http://localhost:8080
    description: Order Service
  - url: http://localhost:8081
    description: Payment Service
  - url: http://localhost:8082
    description: Inventory Service
  - url: http://localhost:8083
    description: Notification Service
...
paths:
...
  /api/orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      description: Creates a new order and triggers the SAGA orchestration
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieves order details by order ID
      operationId: getOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
        - totalAmount
      properties:
        customerId:
          type: string
          example: "CUST-001"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
          example: 299.99
        customerEmail:
          type: string
          format: email
          example: "customer@example.com"
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - price
      properties:
        productId:
          type: string
          example: "PROD-001"
        productName:
          type: string
          example: "Laptop"
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: double
          example: 149.99
    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
          example: "ORD-123456"
        customerId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
        status:
          type: string
          enum:
            - PENDING
            - PAYMENT_PROCESSING
            - PAYMENT_COMPLETED
            - INVENTORY_RESERVING
            - INVENTORY_RESERVED
            - NOTIFYING
            - COMPLETED
            - PAYMENT_FAILED
            - INVENTORY_FAILED
            - CANCELLED
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string